+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to the "translator-helper" extension will be documented in t
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.3.0]
 
-- Initial release
+- Automatically copy the translated text to clipboard for the `translate` command.
+
+## [0.2.0]
+
+- Update dependencies.
+
+## [0.1.0]
+
+- Add reverse translation.
+
+## [0.0.3]
+
+- Initial release.